Experimented with the pixelated border around work cards and tried various ways to make it responsive. This is not 100% resolved.
main.js => rewrote the close modal function to handle any number of cards if I decide to add more in the future.

Robottom Mono font was not loading in properly (even though it is a google font) so I decided to load it in manually.

stars.css => the absolute positioning of the stars was making the page to scroll horizontally, i.e. page is wider than 1vw. I fixed it on desktop by see that it still happens on smaller displays.